feat(web): add lightbox preview for chat images
