+++ web/src/components/AssistantChat/messages/MessageAttachments.tsx
@@ -1,6 +1,7 @@
 import type { AttachmentMetadata } from '@/types/api'
 import { FileIcon } from '@/components/FileIcon'
 import { isImageMimeType } from '@/lib/fileAttachments'
+import { ImagePreview } from '@/components/ImagePreview'
 
 function formatFileSize(bytes: number): string {
     if (bytes < 1024) return `${bytes} B`
@@ -11,18 +12,20 @@ function formatFileSize(bytes: number): string {
 function ImageAttachment(props: { attachment: AttachmentMetadata }) {
     const { attachment } = props
     return (
-        <div className="relative overflow-hidden rounded-lg">
-            <img
-                src={attachment.previewUrl}
-                alt={attachment.filename}
-                className="max-h-48 max-w-full object-contain"
-            />
-            <div className="absolute inset-x-0 bottom-0 bg-gradient-to-t from-black/60 to-transparent px-2 py-1.5">
-                <span className="text-xs text-white/90 line-clamp-1">
-                    {attachment.filename}
-                </span>
-            </div>
-        </div>
+        <ImagePreview
+            src={attachment.previewUrl ?? ''}
+            fileName={attachment.filename}
+            label={attachment.filename}
+            buttonClassName="relative overflow-hidden rounded-lg text-left cursor-zoom-in"
+            imageClassName="max-h-48 max-w-full object-contain"
+            caption={(
+                <div className="absolute inset-x-0 bottom-0 bg-gradient-to-t from-black/60 to-transparent px-2 py-1.5">
+                    <span className="text-xs text-white/90 line-clamp-1">
+                        {attachment.filename}
+                    </span>
+                </div>
+            )}
+        />
     )
 }
